32 bit message

+++ DALIDriver.h
@@ -85,7 +85,7 @@ public:
     */
     int init();
 
-    void attach(mbed::Callback<void(uint8_t)> status_cb);
+    void attach(mbed::Callback<void(uint32_t)> status_cb);
 
     
     /** Send a standard command on the bus
